# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "This project is the rest API of the student branch of University El Bosque.",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:4000",
  "basePath": "/api",
  "tags": [],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/chapters/{chapter_id}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "description": "Endpoint to get one chapter",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/chapters": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "description": "Endpoint to get all chapters",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Chapters"
        ],
        "description": "Endpoint to create a chapter",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Chapter"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Chapters"
        ],
        "description": "Endpoint to update a chapter",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Chapter"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/chaptersinfo/{chapter_info_id}": {
      "get": {
        "tags": [
          "ChaptersInfo"
        ],
        "description": "Endpoint to get one chapter",
        "parameters": [
          {
            "name": "chapter_info_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/chaptersinfo": {
      "get": {
        "tags": [
          "ChaptersInfo"
        ],
        "description": "Endpoint to get all chaptersinfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "ChaptersInfo"
        ],
        "description": "Endpoint to create a chapterinfo",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChapterInfo"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "ChaptersInfo"
        ],
        "description": "Endpoint to update a chaptersinfo",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChapterInfo"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/chaptersmembers/{chaptermember_id}": {
      "get": {
        "tags": [
          "ChaptersMembers"
        ],
        "description": "Endpoint to get one chaptermember",
        "parameters": [
          {
            "name": "chaptermember_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/chaptersmembers": {
      "get": {
        "tags": [
          "ChaptersMembers"
        ],
        "description": "Endpoint to get all chaptersmembers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "ChaptersMembers"
        ],
        "description": "Endpoint to create a chaptermember",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChapterMember"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "ChaptersMembers"
        ],
        "description": "Endpoint to update a chaptermember",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChapterMember"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/events/{event_id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "description": "Endpoint to get one event",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "description": "Endpoint to get all events",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "description": "Endpoint to create a event",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Event"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Events"
        ],
        "description": "Endpoint to update a event",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Event"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/eventtypes/{event_type_id}": {
      "get": {
        "tags": [
          "EventTypes"
        ],
        "description": "Endpoint to get one eventtype",
        "parameters": [
          {
            "name": "event_type_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/eventtypes": {
      "get": {
        "tags": [
          "EventTypes"
        ],
        "description": "Endpoint to get all eventtypes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "EventTypes"
        ],
        "description": "Endpoint to create a eventtype",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventType"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "EventTypes"
        ],
        "description": "Endpoint to update a eventtype",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventType"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/galleries/{gallery_id}": {
      "get": {
        "tags": [
          "Galleries"
        ],
        "description": "Endpoint to get one gallery",
        "parameters": [
          {
            "name": "gallery_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/galleries": {
      "get": {
        "tags": [
          "Galleries"
        ],
        "description": "Endpoint to get all galleries",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Galleries"
        ],
        "description": "Endpoint to create a gallery",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Gallery"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Galleries"
        ],
        "description": "Endpoint to update a gallery",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Gallery"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/members/{member_id}": {
      "get": {
        "tags": [
          "Members"
        ],
        "description": "Endpoint to get one member",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/members": {
      "get": {
        "tags": [
          "Members"
        ],
        "description": "Endpoint to get all members",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Members"
        ],
        "description": "Endpoint to create a member",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Member"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Members"
        ],
        "description": "Endpoint to update a member",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Member"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/roles/{role_id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "description": "Endpoint to get one role",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "description": "Endpoint to get all roles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "description": "Endpoint to create a role",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Role"
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "description": "Endpoint to update a role",
        "parameters": [
          {
            "name": "body",
            "description": "description",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Role"
            },
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "definitions": {
    "Chapter": {
      "type": "object",
      "properties": {
        "chapter_name": {
          "type": "string",
          "example": ""
        },
        "description": {
          "type": "string",
          "example": ""
        },
        "logo_path": {
          "type": "string",
          "example": ""
        },
        "start_date": {
          "type": "string",
          "example": ""
        },
        "end_date": {
          "type": "string",
          "example": ""
        },
        "isActive": {
          "type": "string",
          "example": ""
        }
      }
    },
    "ChapterInfo": {
      "type": "object",
      "properties": {
        "mission": {
          "type": "string",
          "example": ""
        },
        "vission": {
          "type": "string",
          "example": ""
        },
        "objectives": {
          "type": "string",
          "example": ""
        },
        "chapter_id": {
          "type": "number",
          "example": 0
        }
      }
    },
    "ChapterMember": {
      "type": "object",
      "properties": {
        "role_id": {
          "type": "number",
          "example": 0
        },
        "chapter_id": {
          "type": "number",
          "example": 0
        },
        "member_id": {
          "type": "number",
          "example": 0
        },
        "isActive": {
          "type": "string",
          "example": ""
        },
        "start_date": {
          "type": "string",
          "example": ""
        },
        "end_date": {
          "type": "string",
          "example": ""
        }
      }
    },
    "Event": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": ""
        },
        "description": {
          "type": "string",
          "example": ""
        },
        "start_date": {
          "type": "string",
          "example": ""
        },
        "end_date": {
          "type": "string",
          "example": ""
        },
        "event_type_id": {
          "type": "number",
          "example": 0
        },
        "chapter_id": {
          "type": "number",
          "example": 0
        }
      }
    },
    "EventType": {
      "type": "object",
      "properties": {
        "event_type": {
          "type": "string",
          "example": ""
        }
      }
    },
    "Gallery": {
      "type": "object",
      "properties": {
        "gallery_name": {
          "type": "string",
          "example": ""
        },
        "path": {
          "type": "string",
          "example": ""
        },
        "description": {
          "type": "string",
          "example": ""
        },
        "event_id": {
          "type": "number",
          "example": 0
        }
      }
    },
    "Member": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string",
          "example": ""
        },
        "lastname": {
          "type": "string",
          "example": ""
        },
        "email": {
          "type": "string",
          "example": "email@email.com"
        },
        "password": {
          "type": "string",
          "example": ""
        },
        "document": {
          "type": "string",
          "example": ""
        },
        "phone": {
          "type": "string",
          "example": ""
        },
        "image_path": {
          "type": "string",
          "example": ""
        }
      }
    },
    "Role": {
      "type": "object",
      "properties": {
        "role_name": {
          "type": "string",
          "example": ""
        },
        "description": {
          "type": "string",
          "example": ""
        },
        "canCreate": {
          "type": "string",
          "example": ""
        },
        "canDelete": {
          "type": "string",
          "example": ""
        },
        "canUpdate": {
          "type": "string",
          "example": ""
        }
      }
    }
  }
}